ci: Turn off title validation for single commit PRs (#3313)
<!-- Replace this block with what this PR does and why. Describe what
you'd like reviewers to know, how you applied the engineering
principles, and any interesting tradeoffs made. Delete bullet points
below that don't apply, and update the changelog section as appropriate.
-->

When the semantic commit action was added into the branch, the default
squash commit for a pr with only a single commit was the commit name and
not the PR title. Since then, Github has [added a feature to allow
configuration of the
default](https://github.com/orgs/community/discussions/16271) and we
have that setting set in our repo.

This PR turns off the validation of commits as the check is not valid
anymore.