ci: [1.X] CI failures fixes and disabled tests Jira tracking informations (#3303)
This PR addresses some test failures in new CI definition (present on
develop-2.0.0 branch)

1. **Code Coverage test failure** --> This test was failing because
packed package couldn't be find. This steams from the fact that the path
was wrongly pointing to different place because package pack job was
modified to fix it.
2. **minimalproject standards failure** --> We were focusing before on
making sure that testproject is conforming to standards (which also
means that NGO package conforms to standards) but minimalproject has few
errors which we aim to fix in this PR
3. **testproject-tools-integration standards failure** --> We were
focusing before on making sure that testproject is conforming to
standards (which also means that NGO package conforms to standards) but
testproject-tools-integration has few errors which we aim to fix in this
PR
4. **testproject standards test on 2021 editor** --> it seems to be only
2021 specific. For now this configuration was disabled and added for
tracking
5. **Standalone builds which use combination of ubuntu and 2021 editor**
--> Error that happens is "No usable version of libssl". The temporary
solution is to drop usage of this combination with reasons being that we
still have a coverage with ubuntu on all other editors, we still have
coverage of 2021 editor with win and mac and that the 2021 editor will
soon be out of support
7. **Android test failures on older editors** --> I'm going to
investigate and if those are specific tests (and not infrastructure)
then I will disable them for now since my next step after making this CI
green will be to investigate all disabled tests and fix them.
8. **iOS test failures 2021 specific** --> I'm going to investigate and
if those are specific tests (and not infrastructure) then I will disable
them for now since my next step after making this CI green will be to
investigate all disabled tests and fix them
9. **Consoles build failures on 2023 editor** --> Seems to be specific
to 2023 editor, the errors are for example "Exception: The required GDK
is not installed". I will disable console tests for 2023 editor
temporarily and add it to my list to fix together with other disabled
tests

---------

Co-authored-by: Noel Stephens <noel.stephens@unity3d.com>